+++ NEWS.md
@@ -6,6 +6,9 @@
     to be reported by `estimate()`.
   - `estimate.default`: removed `R`, `null.sim`, `score.deriv` and `folds`
     arguments.
+  - new `c.summary.estimate` S3-method for concatenating `summary.estimate`
+    objects
+>>>>>>> dev
   - sim.default can now operate on function return objects
     of type `estimate.extra` which are constructed via
     `c.estimate(estimate_object, extra_args)`